requirement.txt
Flask==2.2.3
Scrapy==2.8.0
pandas==2.0.0
pandas-gbq==0.19.1
pyyaml==6.0
beautifulsoup4==4.12.2
selenium==4.10.0